+++ P2/include/util.hpp
@@ -20,28 +20,28 @@
 #include <stdlib.h>
 #include <iostream>
 #include <climits>
- #include <stdint.h>
+#include <stdint.h>
 
 
 /* The result type used instead of something unclear like return 0. */
 typedef enum Result { FAIL, SUCCESS } Result; 
 
 /* Free a 2D int array of given width. */
-Result Free2DArray(unsigned** array, unsigned n);
+Result Free2DArray(unsigned long long** array, unsigned long long n);
 
 /* Factorial function. */
-unsigned Fact(unsigned n);
+unsigned long long Fact(unsigned long long n);
 
 /* Print a vector. */
-void print_vector(std::vector<unsigned> v);
+void print_vector(std::vector<unsigned long long> v);
 
 /* Subtract two unsgined ints throwing an exception upon underflow. */
-unsigned subtract(unsigned a, unsigned b);
+unsigned long long subtract(unsigned long long a, unsigned long long b);
 
 /* Return whether a < b <= c. */
-bool between_left(unsigned a, unsigned b, unsigned c);
+bool between_left(unsigned long long a, unsigned long long b, unsigned long long c);
 /* Return whether a <= b < c. */
-bool between_right(unsigned a, unsigned b, unsigned c);
+bool between_right(unsigned long long a, unsigned long long b, unsigned long long c);
