+++ P2/src/parser.cpp
@@ -14,16 +14,16 @@
 using namespace std;
 
 Result ReadRectangleDimensions(fstream &input_file, 
-                               unsigned **rectangle_dims,
-                               unsigned n)
+                               unsigned long long **rectangle_dims,
+                               unsigned long long n)
 {
     string line;
-    unsigned int width, height;
+    unsigned long long width, height;
     /* Skip first line. */
     getline(input_file, line);
 
-    /* Read n lines and fetch two unsigned ints from each of them. */
-    for (unsigned i = 0; i < n; i++) {
+    /* Read n lines and fetch two unsigned long longs from each of them. */
+    for (unsigned long long i = 0; i < n; i++) {
         getline(input_file, line);
         istringstream line_stream(line);
         if (line_stream >> width >> height) {
@@ -37,7 +37,7 @@ Result ReadRectangleDimensions(fstream &input_file,
 }
 
 Result ParseFile(const char *file_name, Instance &inst) {
-    unsigned n;
+    unsigned long long n;
 
     /* Try to open file. */
     fstream input_file(file_name);
@@ -55,10 +55,10 @@ Result ParseFile(const char *file_name, Instance &inst) {
     }
 
     /* Try to read rectangles dims line by line. */
-    unsigned int **rectangle_dims;
-    rectangle_dims = (unsigned**) calloc(n, sizeof(unsigned*));
-    for (unsigned i = 0; i < n; i++)
-        rectangle_dims[i] = (unsigned*) calloc(2, sizeof(unsigned));
+    unsigned long long **rectangle_dims;
+    rectangle_dims = (unsigned long long**) calloc(n, sizeof(unsigned long long*));
+    for (unsigned long long i = 0; i < n; i++)
+        rectangle_dims[i] = (unsigned long long*) calloc(2, sizeof(unsigned long long));
 
     if (ReadRectangleDimensions(input_file, rectangle_dims, n) == FAIL) {
         cout << "ParseFile: Failed to read rectangle dimensions" << endl;
